feat(client): Add management command to generate fake client data
This commit adds a new Django management command that can generate fake
data for the Client model. The command can take two optional arguments:
`account_id` and `quantity`. If `account_id` is provided, the command
will use the account with that ID for the created clients. If not, the
command will let the factory create a new Account for each Client. The
`quantity` argument controls the number of clients to be created and
defaults to 100.

This command will help with generating test data and
can also be useful for load testing, performance testing, etc.